backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.19
groq==0.13.0
chromadb==0.5.23
pypdf2==3.0.1
python-docx==1.1.2
aiofiles==24.1.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
